Set @machine_barcode on initialization
If ruby has warnings turned on, it complains about the use of
uninitialized instance variables. Here we ensure that machine
barcode is always initialized.

Note: We use machine_barcode && machine_barcode.to_i to
maintain compatibility with Ruby 2.2. This probably isn't
something we need to worry about too much in the future.